Add files via upload
Code to be added in source code: <div style="display:none;">The Destroyer</div>
The text would be written on the webpage itself but wont be visible unless selected by the user using cursor or searched for hidden tag in the source code. We uploaded the Drax image as the quote used in it is very relevant to the question but has nothing to do with the question. Flag will be the hidden text. Hide the text somewhere on the edges or in the corner.